+++ modules/registration_module/registration_functions.py
@@ -182,7 +182,7 @@ class EditTransformationMatrix(QWidget):
         self.layer_points.editable = False
         # In the current version of napari (v0.4.17), editable is set to True whenever we change the axis value by clicking on the corresponding slider.
         # This is a quick and dirty hack to force the layer to stay non-editable.
-        self.layer_points.events.editable.connect(self.layer_points_editable)
+        self.layer_points.events.editable.connect(lambda e: setattr(e.source,'editable',False))
 
         # To allow saving transformation matrix before closing (__del__ is called too late)
         # TODO: replace by proper napari close event once implemented (https://forum.image.sc/t/handle-of-close-event-in-napari/61039)
@@ -190,9 +190,6 @@ class EditTransformationMatrix(QWidget):
 
         self.viewer.dims.events.current_step.connect(self.dims_current_step_changed)
         self.dims_last_step=self.viewer.dims.current_step
-
-    def layer_points_editable(self):
-        self.layer_points.editable = False
 
     def dims_current_step_changed(self,event):
         try: